# GH-43891: [C++][Parquet] Faster reading of FIXED_LEN_BYTE_ARRAY data