@pitrou 2025-06-25
cc @wgtmac @mapleFU