@mapleFU 2025-06-25
General ideas LGTM, it's a bit late in my tz and I will take a careful round tomorrow